requirements.txt
selenium
webdriver-manager
dnspython
duckduckgo-search
beautifulsoup4
requests